split bullrun
