+++ package.json
@@ -29,7 +29,7 @@
     "ci": "node ./tools/ci.js",
     "test": "jest src",
     "start-electron-webdriver": "node tests/docker-electron-webdriver/start.js",
-    "spectron": "jest tests/specs/bullrun.spec.js",
+    "spectron": "jest tests/specs/*.spec.js",
     "generate-screenshots": "PWD=$(pwd); cd docker; docker run -v \"$PWD:/github/workspace\" -w \"/github/workspace\" --rm -it $(docker build -q .)"
   },
   "engines": {
